+++ .github/workflows/ReleaseNotes.md
@@ -1,3 +1,4 @@
+* [Core] Fixed not working removal of event handlers from async events (#1479).
 * [Client] Added support for passing the local endpoint (network card) which should be used (#1311).
 * [Client] Exposed _PackageIdentifier_ in _MqttApplicationMessageReceivedEventArgs_ (thanks to @koepalex, #1466).
 * [Server] Added a new event (_ClientAcknowledgedPublishPacketAsync_) which is fired whenever a client acknowledges a PUBLISH packet (QoS 1 and 2, #487).